refactor(tui): wire TuiContext into run(), move event loop to method (#209)
run() is now 66 lines — setup, channels, event loop dispatch, cleanup.
All event loop logic is in TuiContext::run_event_loop().
tui_context.rs is 1,570 lines (will be split in next commit).

Mechanical transformation: all bare locals → self.field references.
All 9 tests pass. Zero behavioral change.
